sp_agent/requirements.txt
fastapi==0.118.3
google==3.0.0
google-adk==1.19.0
google-genai==1.53.0
graphviz==0.21
grpcio
grpcio-status
litellm==1.80.9
openai==2.8.1
openpyxl==3.1.5
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
uvicorn==0.38.0
